Minimalist solution to degeneracy problem with convex hull calculation:
When calculating visibility, perturb each point by a random vector of magnitude approx 1e-13 times the magnitude of the point. For all conceivable fullerene structures, this has no bearing on the geometry and is equivalent to a symbolic perturbation, but with much simpler implementation. Each point is only temporarily perturbed; the actual geometry is never changed.